feat :토큰 전달 함수 작성
